+++ src/lib.rs
@@ -1,5 +1,9 @@
 #![warn(clippy::all, clippy::pedantic)]
 
+use std::collections::VecDeque;
+use std::collections::HashMap;
+use crate::parser::story_parser::story;
+
 pub mod parser;
 
 #[derive(Debug)]
@@ -8,4 +12,20 @@ pub enum StoryEvent {
     Input(String),
     Pause,
     Clear,
+}
+
+pub struct AppData {
+    pub story: VecDeque<StoryEvent>,
+    pub game_store: HashMap<String, String>,
+}
+impl Default for AppData {
+    fn default() -> AppData {
+        let text = include_str!("../story/entry.story");
+        let story = story(text).unwrap();
+
+        AppData {
+            story: VecDeque::from(story),
+            game_store: HashMap::new(),
+        }
+    }
 }
